refactor: Move StorageFile dependents to separate files
