+++ src/Uno.UWP/Storage/ThumbnailMode.cs
@@ -1,0 +1,11 @@
+﻿#pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
+
+
+namespace Windows.Storage
+{
+	public enum ThumbnailMode
+	{
+		MusicView,
+		VideosView
+	}
+}
